requirements.txt
flask==2.3.3
ultralytics==8.0.196
opencv-python==4.8.1.78
pillow==10.0.0
pandas==2.0.3
openpyxl==3.1.2
numpy==1.24.3
waitress==2.1.2